requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx>=0.25.0
pytest>=7.4.0
aiofiles>=23.0.0

